Revert "Task 2.1: SwiftData Setup (#108)"
This reverts commit 18ed43e9d93569fae68a8f5311c5f6cc7375c634.
