+++ Thriftwood/Core/DI/BaseViewModel.swift
@@ -70,4 +70,24 @@ extension BaseViewModel {
     }
 }
 
-
+/// Base implementation of a ViewModel with common state management
+@MainActor
+@Observable
+class DefaultViewModel: BaseViewModel {
+    var isLoading: Bool = false
+    var error: ThriftwoodError?
+    
+    var hasData: Bool {
+        // Override in subclasses
+        false
+    }
+    
+    func load() async {
+        // Override in subclasses
+        AppLogger.general.warning("load() not implemented in \(String(describing: type(of: self)))")
+    }
+    
+    func reload() async {
+        await load()
+    }
+}
