+++ Thriftwood/Core/Logging/Logger.swift
@@ -1,13 +1,23 @@
 //
-//  isolation.swift
+//  Logger.swift
 //  Thriftwood
 //
-//  Created by Matthias Wallner-Géhri on 04.10.25.
+//  Centralized logging framework using OSLog
 //
-
 
 import Foundation
 import OSLog
+
+/// Logging categories for different subsystems
+enum LogCategory: String {
+    case networking = "Networking"
+    case storage = "Storage"
+    case authentication = "Authentication"
+    case ui = "UI"
+    case services = "Services"
+    case navigation = "Navigation"
+    case general = "General"
+}
 
 /// Centralized logger for Thriftwood application
 ///
