+++ Thriftwood/Core/Navigation/Coordinator.swift
@@ -20,17 +20,17 @@ import SwiftUI
 /// - https://www.hackingwithswift.com/articles/175/advanced-coordinator-pattern-tutorial-ios
 ///
 @MainActor
-protocol CoordinatorProtocol: AnyObject {
+protocol Coordinator: AnyObject {
     /// The type of route this coordinator handles
     associatedtype Route: Hashable
     
     /// Array of child coordinators managed by this coordinator.
     /// Child coordinators should be added when creating sub-flows and removed when finished.
-    var childCoordinators: [any CoordinatorProtocol] { get set }
+    var childCoordinators: [any Coordinator] { get set }
     
     /// Reference to the parent coordinator (if this is a child coordinator).
     /// Use `weak` to avoid retain cycles since parent owns the child.
-    var parent: (any CoordinatorProtocol)? { get set }
+    var parent: (any Coordinator)? { get set }
     
     /// The navigation path used with SwiftUI's NavigationStack.
     /// This replaces UINavigationController from the UIKit version.
@@ -53,12 +53,12 @@ protocol CoordinatorProtocol: AnyObject {
     /// Called when a child coordinator has finished its work.
     /// The parent should remove the child from its childCoordinators array.
     /// - Parameter child: The child coordinator that finished
-    func childDidFinish(_ child: any CoordinatorProtocol)
+    func childDidFinish(_ child: any Coordinator)
 }
 
 // MARK: - Default Implementations
 
-extension CoordinatorProtocol {
+extension Coordinator {
     /// Default implementation for navigating to a route
     func navigate(to route: Route) {
         navigationPath.append(route)
@@ -77,7 +77,7 @@ extension CoordinatorProtocol {
     
     /// Default implementation for cleaning up child coordinators.
     /// Uses identity comparison (===) to find the specific child coordinator.
-    func childDidFinish(_ child: any CoordinatorProtocol) {
+    func childDidFinish(_ child: any Coordinator) {
         for (index, coordinator) in childCoordinators.enumerated() {
             if coordinator === child {
                 childCoordinators.remove(at: index)
