--- Thriftwood/Core/Navigation/Route/AppRoute.swift
@@ -1,15 +1,0 @@
-//
-//  AppRoute.swift
-//  Thriftwood
-//
-//  Created by Matthias Wallner-Géhri on 04.10.25.
-//
-
-
-import Foundation
-
-/// Root-level routes for the entire application
-enum AppRoute: Hashable, Sendable {
-    case onboarding
-    case main
-}
