+++ Thriftwood/Core/Storage/Models/ServiceConfiguration.swift
@@ -32,7 +32,9 @@ final class ServiceConfiguration {
     var isEnabled: Bool
     var host: String
     var authenticationType: AuthenticationType
-    // Note: apiKey, username, and password are stored in Keychain, not SwiftData
+    var apiKey: String
+    var username: String
+    var password: String
     var broadcastAddress: String
     var macAddress: String
     var headers: [String: String]
@@ -44,6 +46,9 @@ final class ServiceConfiguration {
         isEnabled: Bool = false,
         host: String = "",
         authenticationType: AuthenticationType = .apiKey,
+        apiKey: String = "",
+        username: String = "",
+        password: String = "",
         broadcastAddress: String = "",
         macAddress: String = "",
         headers: [String: String] = [:]
@@ -53,6 +58,9 @@ final class ServiceConfiguration {
         self.isEnabled = isEnabled
         self.host = host
         self.authenticationType = authenticationType
+        self.apiKey = apiKey
+        self.username = username
+        self.password = password
         self.broadcastAddress = broadcastAddress
         self.macAddress = macAddress
         self.headers = headers
@@ -60,15 +68,13 @@ final class ServiceConfiguration {
 }
 
 extension ServiceConfiguration {
-    /// Validates the configuration (excluding credentials which are in Keychain)
-    /// - Note: Credential validation should be done separately via KeychainService
     func isValid() -> Bool {
         guard isEnabled else { return true }
         switch serviceType {
         case .radarr, .sonarr, .lidarr, .sabnzbd, .tautulli, .overseerr:
-            return isHostValid()
+            return isHostValid() && !apiKey.isEmpty
         case .nzbget:
-            return isHostValid()
+            return isHostValid() && !username.isEmpty && !password.isEmpty
         case .wakeOnLAN:
             return !broadcastAddress.isEmpty && !macAddress.isEmpty && isMACAddressValid()
         }
